Refactor: Generalize VIDEO_CHAPTERS into VIDEO
